Rename CertifOps poll test file
