+++ libparsec/crates/client/tests/unit/certif/mod.rs
@@ -7,7 +7,7 @@ mod add_sequester_authority_certificate;
 mod add_user_certificate;
 mod add_user_revoked_certificate;
 mod add_user_update_certificate;
-mod poll;
+mod poll_server_for_new_certificates;
 mod store;
 mod utils;
 mod validate_manifest;
